dom/media/test/test_streams_element_capture_reset.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that reloading and seeking in a media element that's being captured doesn't crash</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script src="manifest.js"></script>
</head>
<body>
<video id="v"></video>
<video id="vout"></video>
<video id="vout_untilended"></video>
<pre id="test">
<script>
const v = document.getElementById('v');
const vout = document.getElementById('vout');
const vout_untilended = document.getElementById('vout_untilended');

function dumpEvent(event) {
  const video = event.target;
  info(
    `${video.name}:${video.id} GOT EVENT ${event.type} ` +
    `currentTime=${video.currentTime} paused=${video.paused} ` +
    `ended=${video.ended} readyState=${video.readyState}`
  );
}

function unexpected(event) {
  ok(false, `${event.type} event received on ${event.target.id} unexpectedly`);
};

const events = ["timeupdate", "seeking", "seeked", "ended", "playing", "pause"];
for (const e of events) {
  v.addEventListener(e, dumpEvent);
  vout.addEventListener(e, dumpEvent);
  vout_untilended.addEventListener(e, dumpEvent);
}

function isWithinEps(a, b, msg) {
  ok(Math.abs(a - b) < 0.01,
     "Got " + a + ", expected " + b + "; " + msg);
}

function isGreaterThanOrEqualEps(a, b, msg) {
  ok(a >= b - 0.01,
     "Got " + a + ", expected at least " + b + "; " + msg);
}

async function startTest(test) {
  const seekTime = test.duration/2;

  vout.onended = unexpected;

  v.src = test.name;
  v.name = test.name;
  vout.name = test.name;
  vout_untilended.name = test.name;
  v.preload = "metadata";
  await new Promise(r => v.onloadedmetadata = r);

  vout.srcObject = v.mozCaptureStream();
  vout.play();

  vout_untilended.srcObject = v.mozCaptureStreamUntilEnded();
  vout_untilended.play();

  v.play();

  await Promise.all([
    new Promise(r => v.onended = r),
    new Promise(r => vout_untilended.onended = r),
  ]);

  isGreaterThanOrEqualEps(vout.currentTime, test.duration,
    "checking vout.currentTime at first 'ended' event");
  isGreaterThanOrEqualEps(v.currentTime, test.duration,
    "checking v.currentTime at first 'ended' event");
  ok(!vout.ended, "checking vout has not ended");
  ok(vout_untilended.ended, "checking vout_untilended has actually ended");

  v.pause();
  v.currentTime = seekTime;
  await new Promise(r => v.onseeked = r);

  isGreaterThanOrEqualEps(v.currentTime, seekTime, "Finished seeking");
  isGreaterThanOrEqualEps(vout.currentTime, test.duration,
    "checking vout.currentTime has not changed after seeking");

  vout_untilended.onplaying = unexpected;
  vout_untilended.onended = unexpected;
  v.play();

  await new Promise(r => v.onended = r);

  isGreaterThanOrEqualEps(v.currentTime, test.duration,
    "checking v.currentTime at second 'ended' event");
  isGreaterThanOrEqualEps(vout.currentTime,
    (test.duration - seekTime) + test.duration,
    "checking vout.currentTime after seeking and playing through again");
  v.src = test.name + "?1";
  v.play();

  await new Promise(r => v.onended = r);

  isGreaterThanOrEqualEps(v.currentTime, test.duration,
    "checking v.currentTime at third 'ended' event");
  isGreaterThanOrEqualEps(vout.currentTime,
    (test.duration - seekTime) + test.duration*2,
    "checking vout.currentTime after seeking, playing through and reloading");
}

(async () => {
  SimpleTest.waitForExplicitFinish();
  try {
    const testVideo = getPlayableVideo(gSmallTests);
    if (testVideo) {
      await startTest(testVideo);
    } else {
      todo(false, "No playable video");
    }
  } catch(e) {
    ok(false, `Error: ${e}`);
  } finally {
    SimpleTest.finish();
  }
})();
</script>
</pre>
</body>
</html>
